Rename template files to better match convention
This PR renames "root.js" to "App.js", so it makes more sense what the file contains.

Also renames "mainTest.js" to "App.test.js", to better match Jest convention.
