+++ react-native-scripts/template/App.js
