+++ react-native-scripts/template/__tests__/App.test.js
